punchd/serializers.py
from django.contrib.auth.models import User
from .models import Business, Offer, OfferInstance, Punch
from rest_framework import serializers

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('url', 'username', 'email', 'is_staff')

class BusinessSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Business
        fields = ('id', 'name', 'qrcode')