requirements.txt
Django==1.8.4
gunicorn==19.3.0
whitenoise==2.0.3
flask==0.10.1
#dj-database-url==0.3.0
#SQLAlchemy==1.0.4
djangorestframework=3.2.3